requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Bootstrap-Flask==2.3.3
gunicorn==21.2.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.4
tensorflow-cpu==2.16.1
keras==3.3.3